refactor(android): extract Mapscore state into typed sessions
