+++ android/src/main/kotlin/hamza/dali/flutter_osm_plugin/mapscore/LegacyMethodChannelAdapter.kt
@@ -1,25 +1,88 @@
 package hamza.dali.flutter_osm_plugin.mapscore
 
 import io.flutter.plugin.common.BinaryMessenger
+import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 
+/** Existing Android command names, kept byte-for-byte compatible with Dart. */
+internal enum class LegacyMapCommand(val methodName: String) {
+    CHANGE_TILE("change#tile"),
+    USE_VISIBILITY_INFO_WINDOW("use#visiblityInfoWindow"),
+    CONFIG_ZOOM("config#Zoom"),
+    ZOOM("Zoom"),
+    GET_ZOOM("get#Zoom"),
+    CHANGE_STEP_ZOOM("change#stepZoom"),
+    ZOOM_TO_REGION("zoomToRegion"),
+    SHOW_ZOOM_CONTROLLER("showZoomController"),
+    CURRENT_LOCATION("currentLocation"),
+    INIT_MAP("initMap"),
+    LIMIT_AREA("limitArea"),
+    REMOVE_LIMIT_AREA("remove#limitArea"),
+    CHANGE_POSITION("changePosition"),
+    TRACK_ME("trackMe"),
+    DEACTIVATE_TRACK_ME("deactivateTrackMe"),
+    START_LOCATION_UPDATING("startLocationUpdating"),
+    STOP_LOCATION_UPDATING("stopLocationUpdating"),
+    MAP_CENTER("map#center"),
+    MAP_BOUNDS("map#bounds"),
+    USER_POSITION("user#position"),
+    MOVE_TO_POSITION("moveTo#position"),
+    USER_REMOVE_MARKER_POSITION("user#removeMarkerPosition"),
+    DELETE_ROAD("delete#road"),
+    DRAW_MULTI_ROAD("draw#multi#road"),
+    CLEAR_ROADS("clear#roads"),
+    MARKER_ICON("marker#icon"),
+    DRAW_ROAD_MANUALLY("drawRoad#manually"),
+    STATIC_POSITION("staticPosition"),
+    STATIC_POSITION_ICON_MARKER("staticPosition#IconMarker"),
+    DRAW_CIRCLE("draw#circle"),
+    DRAW_RECT("draw#rect"),
+    REMOVE_CIRCLE("remove#circle"),
+    REMOVE_RECT("remove#rect"),
+    CLEAR_SHAPES("clear#shapes"),
+    MAP_ORIENTATION("map#orientation"),
+    USER_LOCATION_MARKERS("user#locationMarkers"),
+    ADD_MARKER("add#Marker"),
+    UPDATE_MARKER("update#Marker"),
+    CHANGE_MARKER("change#Marker"),
+    GET_GEOPOINTS("get#geopoints"),
+    DELETE_MARKERS("delete#markers"),
+    TOGGLE_ALL_LAYER("toggle#Alllayer"),
+    ;
+
+    companion object {
+        private val byMethodName = values().associateBy(LegacyMapCommand::methodName)
+
+        fun fromMethodName(methodName: String): LegacyMapCommand? = byMethodName[methodName]
+    }
+}
+
 /**
- * Compatibility transport for the existing Android method-channel contract.
+ * Compatibility transport for the existing Android MethodChannel contract.
  *
- * Command dispatch and event emission share this object, while the map view
- * remains responsible for implementing the legacy command handler. A future
- * JNI transport can be attached to the same map session without changing the
- * renderer's event callbacks.
+ * This adapter is the only component that decodes legacy string method names.
+ * It forwards a typed command to the shared map session and also carries
+ * native-to-Dart events for that session.
  */
-class LegacyMethodChannelAdapter(
+internal class LegacyMethodChannelAdapter(
     messenger: BinaryMessenger,
     viewId: Int,
-) : MapEventSink {
+    private val dispatch: (LegacyMapCommand, MethodCall, MethodChannel.Result) -> Unit,
+) : MapEventSink, MethodChannel.MethodCallHandler {
     private val channel = MethodChannel(messenger, "plugins.dali.hamza/osmview_$viewId")
     private val eventSink = MethodChannelMapEventSink(channel)
 
-    fun setMethodCallHandler(handler: MethodChannel.MethodCallHandler?) {
-        channel.setMethodCallHandler(handler)
+    init {
+        channel.setMethodCallHandler(this)
+    }
+
+    override fun onMethodCall(call: MethodCall, result: MethodChannel.Result) {
+        val command = LegacyMapCommand.fromMethodName(call.method)
+        if (command == null) {
+            result.notImplemented()
+            return
+        }
+        dispatch(command, call, result)
     }
 
     override fun emit(method: String, arguments: Any?) {
